restructure directory and rename
