+++ src/dashboard/components/pages/questioncategories/ExamSubTypePage.jsx
@@ -1,21 +1,21 @@
-import QuestionCategoryFormWithSelect from "@/dashboard/components/organism/QuestionCategoryFormWithSelect";
-import useDataTableColumns from "@/dashboard/hooks/useDataTableColumns";
+import useDataTableColumns from "@/dashboard/components/molecules/categories/useDataTableColumns";
 import { useGetQuestionsCategoryQuery } from "@/features/questions/questionsCategoryApi";
-import Loading from "../../../atoms/Loading";
-import PageTitle from "../../../atoms/PageTitle";
-import ThemeSwitch from "../../../atoms/ThemeSwitch";
-import UserNav from "../../../organism/UserNav";
-import { DataTable } from "../../../templates/DataTable";
-import { Layout } from "../../../templates/Layout";
+import Loading from "../../atoms/Loading";
+import PageTitle from "../../atoms/PageTitle";
+import ThemeSwitch from "../../atoms/ThemeSwitch";
+import QuestionCategoryFormWithSelect from "../../organism/QuestionCategoryFormWithSelect";
+import UserNav from "../../organism/UserNav";
+import { DataTable } from "../../templates/DataTable";
+import { Layout } from "../../templates/Layout";
 
-const TopicsForQuestionPage = () => {
-    const { data: topicsData, isLoading, isSuccess, refetch } = useGetQuestionsCategoryQuery("topics");
-    const columns = useDataTableColumns("topics");
+export default function ExamSubTypePage() {
+    const { data: examSubtypeData, isLoading, isSuccess, refetch } = useGetQuestionsCategoryQuery("exam-sub-types");
+    const columns = useDataTableColumns("exam-sub-types");
 
     return (
         <Layout>
             <Layout.Header>
-                <PageTitle title={"Topics"} />
+                <PageTitle title={"Exam sub-type"} />
                 <div className='ml-auto flex items-center space-x-4'>
                     <ThemeSwitch />
                     <UserNav />
@@ -24,23 +24,23 @@ const TopicsForQuestionPage = () => {
 
             <Layout.Body>
                 <QuestionCategoryFormWithSelect
-                    type={"topics"}
+                    type={"exam-sub-types"}
                     refetchOnQuestionsCategoryQuery={refetch}
-                    fromTopics={true}
+                    fromExamSubTypes={true}
                 />
 
                 <div className='mt-8 mb-2 flex items-center justify-between space-y-2'>
                     <div>
                         <h2 className='text-2xl font-bold tracking-tight'>Welcome back!</h2>
                         <p className='text-muted-foreground'>
-                            Here&apos;s a list of exam section group!
+                            Here&apos;s a list of exam sub-types!
                         </p>
                     </div>
                 </div>
                 <div className='-mx-4 flex-1 overflow-auto px-4 py-1 lg:flex-row lg:space-x-12 lg:space-y-0'>
                     {
                         isLoading ? <Loading /> : (
-                            isSuccess && topicsData?.data?.data && <DataTable data={topicsData?.data?.data} columns={columns} />
+                            isSuccess && examSubtypeData?.data?.data && <DataTable data={examSubtypeData?.data?.data} columns={columns} />
                         )
                     }
                 </div>
@@ -48,5 +48,3 @@ const TopicsForQuestionPage = () => {
         </Layout>
     )
 }
-
-export default TopicsForQuestionPage
