+++ src/dashboard/components/pages/questioncategories/LevelPage.jsx
@@ -1,21 +1,21 @@
-import useDataTableColumns from "@/dashboard/hooks/useDataTableColumns";
+import Loading from "@/dashboard/components/atoms/Loading";
+import useDataTableColumns from "@/dashboard/components/molecules/categories/useDataTableColumns";
+import QuestionCategoryFormWithSelect from "@/dashboard/components/organism/QuestionCategoryFormWithSelect";
+import { DataTable } from "@/dashboard/components/templates/DataTable";
 import { useGetQuestionsCategoryQuery } from "@/features/questions/questionsCategoryApi";
-import Loading from "../../../atoms/Loading";
-import PageTitle from "../../../atoms/PageTitle";
-import ThemeSwitch from "../../../atoms/ThemeSwitch";
-import QuestionCategoryFormWithSelect from "../../../organism/QuestionCategoryFormWithSelect";
-import UserNav from "../../../organism/UserNav";
-import { DataTable } from "../../../templates/DataTable";
-import { Layout } from "../../../templates/Layout";
+import PageTitle from "../../atoms/PageTitle";
+import ThemeSwitch from "../../atoms/ThemeSwitch";
+import UserNav from "../../organism/UserNav";
+import { Layout } from "../../templates/Layout";
 
-const ExamSubTypeForQuestionPage = () => {
-    const { data: examSubtypeData, isLoading, isSuccess, refetch } = useGetQuestionsCategoryQuery("exam-sub-types");
-    const columns = useDataTableColumns("exam-sub-types");
+export default function LevelPage() {
+    const { data: levelData, isLoading, isSuccess, refetch } = useGetQuestionsCategoryQuery("levels");
+    const columns = useDataTableColumns("levels");
 
     return (
         <Layout>
             <Layout.Header>
-                <PageTitle title={"Exam sub-type"} />
+                <PageTitle title={"Level"} />
                 <div className='ml-auto flex items-center space-x-4'>
                     <ThemeSwitch />
                     <UserNav />
@@ -24,23 +24,23 @@ const ExamSubTypeForQuestionPage = () => {
 
             <Layout.Body>
                 <QuestionCategoryFormWithSelect
-                    type={"exam-sub-types"}
+                    fromLevels={true}
+                    type={"levels"}
                     refetchOnQuestionsCategoryQuery={refetch}
-                    fromExamSubTypes={true}
                 />
 
                 <div className='mt-8 mb-2 flex items-center justify-between space-y-2'>
                     <div>
                         <h2 className='text-2xl font-bold tracking-tight'>Welcome back!</h2>
                         <p className='text-muted-foreground'>
-                            Here&apos;s a list of exam sub-types!
+                            Here&apos;s a list of exam section group!
                         </p>
                     </div>
                 </div>
                 <div className='-mx-4 flex-1 overflow-auto px-4 py-1 lg:flex-row lg:space-x-12 lg:space-y-0'>
                     {
                         isLoading ? <Loading /> : (
-                            isSuccess && examSubtypeData?.data?.data && <DataTable data={examSubtypeData?.data?.data} columns={columns} />
+                            isSuccess && levelData?.data?.data && <DataTable data={levelData?.data?.data} columns={columns} />
                         )
                     }
                 </div>
@@ -48,4 +48,3 @@ const ExamSubTypeForQuestionPage = () => {
         </Layout>
     )
 }
-export default ExamSubTypeForQuestionPage
