+++ src/dashboard/components/pages/questioncategories/SubTopicsPage.jsx
@@ -1,21 +1,21 @@
-import Loading from "@/dashboard/components/atoms/Loading";
-import QuestionCategoryForm from "@/dashboard/components/organism/QuestionCategoryForm";
-import { DataTable } from "@/dashboard/components/templates/DataTable";
-import useDataTableColumns from "@/dashboard/hooks/useDataTableColumns";
+import useDataTableColumns from "@/dashboard/components/molecules/categories/useDataTableColumns";
+import QuestionCategoryFormWithSelect from "@/dashboard/components/organism/QuestionCategoryFormWithSelect";
 import { useGetQuestionsCategoryQuery } from "@/features/questions/questionsCategoryApi";
-import PageTitle from "../../../atoms/PageTitle";
-import ThemeSwitch from "../../../atoms/ThemeSwitch";
-import UserNav from "../../../organism/UserNav";
-import { Layout } from "../../../templates/Layout";
+import Loading from "../../atoms/Loading";
+import PageTitle from "../../atoms/PageTitle";
+import ThemeSwitch from "../../atoms/ThemeSwitch";
+import UserNav from "../../organism/UserNav";
+import { DataTable } from "../../templates/DataTable";
+import { Layout } from "../../templates/Layout";
 
-const YearForQuestionPage = () => {
-    const { data: yearData, isLoading, isSuccess, refetch } = useGetQuestionsCategoryQuery("years");
-    const columns = useDataTableColumns("years");
+export default function SubTopicsPage() {
+    const { data: subtopicsData, isLoading, isSuccess, refetch } = useGetQuestionsCategoryQuery("sub-topics");
+    const columns = useDataTableColumns("sub-topics");
 
     return (
         <Layout>
             <Layout.Header>
-                <PageTitle title={"Year"} />
+                <PageTitle title={"Sub Topics"} />
                 <div className='ml-auto flex items-center space-x-4'>
                     <ThemeSwitch />
                     <UserNav />
@@ -23,23 +23,24 @@ const YearForQuestionPage = () => {
             </Layout.Header>
 
             <Layout.Body>
-                <QuestionCategoryForm
-                    type={"years"}
+                <QuestionCategoryFormWithSelect
+                    type={"sub-topics"}
                     refetchOnQuestionsCategoryQuery={refetch}
+                    fromSubtopics={true}
                 />
 
                 <div className='mt-8 mb-2 flex items-center justify-between space-y-2'>
                     <div>
                         <h2 className='text-2xl font-bold tracking-tight'>Welcome back!</h2>
                         <p className='text-muted-foreground'>
-                            Here&apos;s a list of exam section year!
+                            Here&apos;s a list of exam section group!
                         </p>
                     </div>
                 </div>
                 <div className='-mx-4 flex-1 overflow-auto px-4 py-1 lg:flex-row lg:space-x-12 lg:space-y-0'>
                     {
                         isLoading ? <Loading /> : (
-                            isSuccess && yearData?.data?.data && <DataTable data={yearData?.data?.data} columns={columns} />
+                            isSuccess && subtopicsData?.data?.data && <DataTable data={subtopicsData?.data?.data} columns={columns} />
                         )
                     }
                 </div>
@@ -47,4 +48,3 @@ const YearForQuestionPage = () => {
         </Layout>
     )
 }
-export default YearForQuestionPage
