+++ src/dashboard/components/pages/questioncategories/SubjectPage.jsx
@@ -1,21 +1,21 @@
 import Loading from "@/dashboard/components/atoms/Loading";
+import useDataTableColumns from "@/dashboard/components/molecules/categories/useDataTableColumns";
 import QuestionCategoryFormWithSelect from "@/dashboard/components/organism/QuestionCategoryFormWithSelect";
 import { DataTable } from "@/dashboard/components/templates/DataTable";
-import useDataTableColumns from "@/dashboard/hooks/useDataTableColumns";
 import { useGetQuestionsCategoryQuery } from "@/features/questions/questionsCategoryApi";
-import PageTitle from "../../../atoms/PageTitle";
-import ThemeSwitch from "../../../atoms/ThemeSwitch";
-import UserNav from "../../../organism/UserNav";
-import { Layout } from "../../../templates/Layout";
+import PageTitle from "../../atoms/PageTitle";
+import ThemeSwitch from "../../atoms/ThemeSwitch";
+import UserNav from "../../organism/UserNav";
+import { Layout } from "../../templates/Layout";
 
-const LevelForQuestionPage = () => {
-    const { data: levelData, isLoading, isSuccess, refetch } = useGetQuestionsCategoryQuery("levels");
-    const columns = useDataTableColumns("levels");
+export default function SubjectPage() {
+    const { data: subjectData, isLoading, isSuccess, refetch } = useGetQuestionsCategoryQuery("subjects");
+    const columns = useDataTableColumns("subjects");
 
     return (
         <Layout>
             <Layout.Header>
-                <PageTitle title={"Level"} />
+                <PageTitle title={"Subject"} />
                 <div className='ml-auto flex items-center space-x-4'>
                     <ThemeSwitch />
                     <UserNav />
@@ -24,12 +24,12 @@ const LevelForQuestionPage = () => {
 
             <Layout.Body>
                 <QuestionCategoryFormWithSelect
-                    fromLevels={true}
-                    type={"levels"}
+                    type={"subjects"}
                     refetchOnQuestionsCategoryQuery={refetch}
+                    fromSubjects={true}
                 />
 
-                <div className='mt-8 mb-2 flex items-center justify-between space-y-2'>
+                <div className='mt-8 mb-2 flex items-center justify-between space-y-2 '>
                     <div>
                         <h2 className='text-2xl font-bold tracking-tight'>Welcome back!</h2>
                         <p className='text-muted-foreground'>
@@ -40,7 +40,7 @@ const LevelForQuestionPage = () => {
                 <div className='-mx-4 flex-1 overflow-auto px-4 py-1 lg:flex-row lg:space-x-12 lg:space-y-0'>
                     {
                         isLoading ? <Loading /> : (
-                            isSuccess && levelData?.data?.data && <DataTable data={levelData?.data?.data} columns={columns} />
+                            isSuccess && subjectData?.data?.data && <DataTable data={subjectData?.data?.data} columns={columns} />
                         )
                     }
                 </div>
@@ -48,5 +48,3 @@ const LevelForQuestionPage = () => {
         </Layout>
     )
 }
-
-export default LevelForQuestionPage
