+++ src/dashboard/components/pages/questioncategories/YearPage.jsx
@@ -1,21 +1,21 @@
 import Loading from "@/dashboard/components/atoms/Loading";
-import QuestionCategoryFormWithSelect from "@/dashboard/components/organism/QuestionCategoryFormWithSelect";
+import useDataTableColumns from "@/dashboard/components/molecules/categories/useDataTableColumns";
+import QuestionCategoryForm from "@/dashboard/components/organism/QuestionCategoryForm";
 import { DataTable } from "@/dashboard/components/templates/DataTable";
-import useDataTableColumns from "@/dashboard/hooks/useDataTableColumns";
 import { useGetQuestionsCategoryQuery } from "@/features/questions/questionsCategoryApi";
-import PageTitle from "../../../atoms/PageTitle";
-import ThemeSwitch from "../../../atoms/ThemeSwitch";
-import UserNav from "../../../organism/UserNav";
-import { Layout } from "../../../templates/Layout";
+import PageTitle from "../../atoms/PageTitle";
+import ThemeSwitch from "../../atoms/ThemeSwitch";
+import UserNav from "../../organism/UserNav";
+import { Layout } from "../../templates/Layout";
 
-const SubjectForQuestionPage = () => {
-    const { data: subjectData, isLoading, isSuccess, refetch } = useGetQuestionsCategoryQuery("subjects");
-    const columns = useDataTableColumns("subjects");
+export default function YearPage() {
+    const { data: yearData, isLoading, isSuccess, refetch } = useGetQuestionsCategoryQuery("years");
+    const columns = useDataTableColumns("years");
 
     return (
         <Layout>
             <Layout.Header>
-                <PageTitle title={"Subject"} />
+                <PageTitle title={"Year"} />
                 <div className='ml-auto flex items-center space-x-4'>
                     <ThemeSwitch />
                     <UserNav />
@@ -23,24 +23,23 @@ const SubjectForQuestionPage = () => {
             </Layout.Header>
 
             <Layout.Body>
-                <QuestionCategoryFormWithSelect
-                    type={"subjects"}
+                <QuestionCategoryForm
+                    type={"years"}
                     refetchOnQuestionsCategoryQuery={refetch}
-                    fromSubjects={true}
                 />
 
-                <div className='mt-8 mb-2 flex items-center justify-between space-y-2 '>
+                <div className='mt-8 mb-2 flex items-center justify-between space-y-2'>
                     <div>
                         <h2 className='text-2xl font-bold tracking-tight'>Welcome back!</h2>
                         <p className='text-muted-foreground'>
-                            Here&apos;s a list of exam section group!
+                            Here&apos;s a list of exam section year!
                         </p>
                     </div>
                 </div>
                 <div className='-mx-4 flex-1 overflow-auto px-4 py-1 lg:flex-row lg:space-x-12 lg:space-y-0'>
                     {
                         isLoading ? <Loading /> : (
-                            isSuccess && subjectData?.data?.data && <DataTable data={subjectData?.data?.data} columns={columns} />
+                            isSuccess && yearData?.data?.data && <DataTable data={yearData?.data?.data} columns={columns} />
                         )
                     }
                 </div>
@@ -48,5 +47,3 @@ const SubjectForQuestionPage = () => {
         </Layout>
     )
 }
-
-export default SubjectForQuestionPage
